semantic-net/pyproject.toml
[project]
name = "semantic-net"
version = "0.1.0"
description = "Federated semantic communication experiments for Jetson-class devices."
readme = "README.md"
requires-python = ">=3.9"
license = "MIT"
authors = [{ name = "Kohsuke Yamakawa" }]
dependencies = [
    "torch>=2.1",
    "numpy>=1.24,<2.0",
    "pyyaml>=6.0",
    "flwr>=1.7",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
]

[tool.setuptools]
package-dir = {"" = "src"}
py-modules = ["run_single", "run_fed_client", "run_fed_server"]

[tool.setuptools.packages.find]
where = ["src"]

[project.scripts]
semantic-run-single = "run_single:main"
semantic-run-fed-client = "run_fed_client:main"
semantic-run-fed-server = "run_fed_server:main"

[build-system]
requires = ["setuptools>=65", "wheel"]
build-backend = "setuptools.build_meta"

[tool.uv]
dev-dependencies = [
    "pytest>=7.4",
]
